Inline BitLinear, drop trigram_side_memory dead code, fix byte accounting
The previous d4a2208 fix (bundle local modules/) made the submission
runnable but introduced a byte-accounting hole: the harness reported
code_bytes=104,676 (just train_gpt.py) while the artifact actually
shipped 146,844 bytes of code (train_gpt.py + 3,674-byte
modules/bitlinear.py + 38,494-byte modules/trigram_side_memory.py). The
upstream rule is "All counted code should live in the train_gpt.py
script."

This commit makes the submission single-file and re-aligns the numbers:

  - Inline BitLinear + pack_ternary + unpack_ternary into train_gpt.py
    near line 705 (the original BitLinear import site). All four import
    sites that previously read `from modules.bitlinear import ...`
    (lines 502, 575, 705, 2178) now resolve in-module.

  - Delete modules/trigram_side_memory.py entirely. It was 38,494 bytes
    of dead code under default TRIGRAM_SIDE_MEMORY=0; all three lazy
    imports inside train_gpt.py were behind conditional guards that
    don't fire under this submission's config. The three import sites
    (lines 1482, 1519, 2000) now raise NotImplementedError so a future
    reviewer who flips the flag gets a clear error instead of a silent
    ModuleNotFoundError.

  - env.sh switches back to the repo-root run convention (no more
    DATA_PATH=../../../...). DATA_PATH/TOKENIZER_PATH defaults already
    resolve from repo root, matching every other records-folder
    submission.

  - README "Command", "Key metrics", "Files" sections rewritten:
      code_bytes:   104,676 -> 106,722  (single-file train_gpt.py)
      payload:      11,969,746          (unchanged; same trained checkpoint)
      total:        12,074,422 -> 12,076,468  (~12.08 MB)
      headroom:     ~3.92 MB under 16 MB cap

  - README "Comparison" + submission.json comparison_baseline now
    explicitly call out the naive records-track baseline (1.2244) and
    note +0.076 BPB worse despite ~9x compute. Previous wording called
    1.1063 "the records baseline" which was wrong (it's a mid-tier
    PR #1204 entry).

  - submission.json adds payload_bytes alongside code_bytes/bytes_total
    so the accounting is self-explanatory; notes field acknowledges the
    train_seed1337.log carries the pre-cleanup numbers from the
    original three-file run while the shipped artifact is single-file.

Trained model checkpoint unchanged — only code-side accounting moved.
