+++ records/track_non_record_16mb/2026-04-30_KillMamba2_TriParallel_n7_Ternary_EMA_4xH200_1hr/README.md
@@ -1,6 +1,6 @@
 First SSM-based entry in either track. Trained on 4×H200 SXM for one hour rather than 8×H100 for ten minutes — non-record on both hardware and time. Eval is the standard root-harness full-window val + int8/brotli quant; no TTT, no sliding-window, no GPTQ.
 
-**val_bpb 1.30040 / 12.07 MB / seed=1337**
+**val_bpb 1.30040 / 12.08 MB / seed=1337**
 
 ## Architecture
 
@@ -41,15 +41,15 @@ First SSM-based entry in either track. Trained on 4×H200 SXM for one hour rathe
 
 ## Command
 
-`train_gpt.py` imports a small local `modules/` package (`modules.bitlinear` for the BitNet-b1.58 ternary path; `modules.trigram_side_memory` is referenced under `TRIGRAM_SIDE_MEMORY=1`, which is **off** here but kept so the script remains import-clean). To make `modules/` resolvable, run from **inside this folder** rather than from the repo root:
+`train_gpt.py` is fully self-contained — `BitLinear` plus `pack_ternary` / `unpack_ternary` are inlined directly into the script (no local helper modules). Run from the **repo root**, same convention as every other records-folder submission:
 
 ```bash
-cd records/track_non_record_16mb/2026-04-30_KillMamba2_TriParallel_n7_Ternary_EMA_4xH200_1hr/
-source ./env.sh
-torchrun --standalone --nproc_per_node=4 train_gpt.py
+source records/track_non_record_16mb/2026-04-30_KillMamba2_TriParallel_n7_Ternary_EMA_4xH200_1hr/env.sh
+torchrun --standalone --nproc_per_node=4 \
+  records/track_non_record_16mb/2026-04-30_KillMamba2_TriParallel_n7_Ternary_EMA_4xH200_1hr/train_gpt.py
 ```
 
-`env.sh` sets `DATA_PATH=../../../data/datasets/fineweb10B_sp1024` and `TOKENIZER_PATH=../../../data/tokenizers/fineweb_1024_bpe.model` (three levels up from this folder reaches the repo root). It also sets `CONTROL_TENSOR_NAME_PATTERNS` (load-bearing — keeps SSM dynamics buffers fp32 under ternary quantization). `MAX_WALLCLOCK_SECONDS=3600` is the binding cap; `ITERATIONS=20000` is just an upper bound.
+`env.sh` sets `CONTROL_TENSOR_NAME_PATTERNS` (load-bearing — keeps SSM dynamics buffers fp32 under ternary quantization), the topology / quant / EMA / optimizer knobs, and the eval cadence. `DATA_PATH` and `TOKENIZER_PATH` are not exported because their script-side defaults (`./data/datasets/fineweb10B_sp1024` and `./data/tokenizers/fineweb_1024_bpe.model`) already resolve correctly from the repo root. `MAX_WALLCLOCK_SECONDS=3600` is the binding cap; `ITERATIONS=20000` is just an upper bound.
 
 ## Key metrics
 
@@ -60,23 +60,25 @@ torchrun --standalone --nproc_per_node=4 train_gpt.py
 - Step time: `821.94 ms`
 - Steps: `4,380`
 - Tokens trained: `4,380 × 524,288 ≈ 2.30B`
-- Artifact (int8 + ternary-packed + brotli): `12,074,422 bytes`
-- Code size: `104,676 bytes`
+- Code size: `106,722 bytes` (`train_gpt.py`; single-file, no local helper modules)
+- Compressed model payload (int8 + 2-bit-packed ternary + brotli q=11): `11,969,746 bytes`
+- **Artifact total: `12,076,468 bytes` = code + payload (≈12.08 MB; 16 MB cap honored with ~3.92 MB headroom)**
 - Model parameters: `61,657,752`
 - Hardware: 4×H200 SXM (141GB HBM3e per GPU), `--nproc 4`, grad_accum=2 (peak GPU memory at run: 114,125 MiB allocated)
 
+The train log (`train_seed1337.log`) shows `code_bytes:104676` and `Total submission size int8+zlib:12074422` because it was generated by the original three-file run (`train_gpt.py` + `modules/bitlinear.py` + `modules/trigram_side_memory.py`). The shipped artifact is now single-file (`bitlinear.py` inlined into `train_gpt.py`; `trigram_side_memory.py` deleted as gated dead code under default `TRIGRAM_SIDE_MEMORY=0`), so the post-cleanup byte counts above replace the log values. The trained model checkpoint is unchanged — only the code-side accounting moved.
+
 ## Comparison
 
-- vs `track_10min_16mb` records baseline `2026-03-31_ParallelResiduals_MiniDepthRecurrence` (1.1063): +0.194 BPB
+- vs `track_10min_16mb` naive baseline (1.2244, 9L 512d sp1024 GQA-4 tied-emb, 8×H100 10 min): **+0.076 BPB worse** despite ~9× the compute budget. This submission is below the naive records-track baseline; it earns its place on the non-record list as the **first SSM entry** rather than as a frontier number.
+- vs `track_10min_16mb` mid-tier records-track frontier `2026-03-31_ParallelResiduals_MiniDepthRecurrence` (1.1063): +0.194 BPB. Most of this gap is duration + missing standard-stack ports (parallel-residuals, sliding-window eval, GPTQ), not architecture.
 - vs `track_non_record_16mb/2026-03-24_106M_Binary_Asymmetric_UNet_FP8_15L...` (1.1239, 8×H100, 2.15h, 8192 BPE): +0.177 BPB
 - vs `track_non_record_16mb/2026-03-18_Quasi10Bfrom50B_SP1024_9x512_KV4_4h_pgut3` (1.2074, 8×H100, 4h): +0.093 BPB
 
 ## Files
 
-- `train_gpt.py` — code snapshot (104,676 bytes — what the harness reports as `code_bytes`)
-- `modules/bitlinear.py` — `BitLinear` (BitNet-b1.58 absmean STE) plus `pack_ternary` / `unpack_ternary` for the 2-bit packed export path; load-bearing under `TERNARY_BODY=1`
-- `modules/trigram_side_memory.py` — trigram side-memory blend; **inert** under this run's `TRIGRAM_SIDE_MEMORY=0` but referenced by import-time guards in `train_gpt.py`, so kept for completeness
-- `env.sh` — canonical environment; source from inside this folder
-- `train_seed1337.log` — training log (partial: pod was stopped before the full `run.log` synced from `/workspace`; the lines preserved cover the headline numbers — pre/post-quant val_bpb, artifact bytes, step times, peak memory, EMA shadow swap)
-- `result.json`, `submission.json` — leaderboard metadata
-- `requirements.txt` — `brotli` and `sentencepiece` are required at quant-export
+- `train_gpt.py` — single-file submission script (106,722 bytes). `BitLinear` plus `pack_ternary` / `unpack_ternary` are inlined; the (gated) `trigram_side_memory` import sites raise `NotImplementedError` so the dead branches are visible-and-unreachable rather than smuggling in a hidden dependency.
+- `env.sh` — canonical environment; source from the repo root.
+- `train_seed1337.log` — training log (partial: pod was stopped before the full `run.log` synced from `/workspace`; the lines preserved cover the headline numbers — pre/post-quant val_bpb, payload bytes, step times, peak memory, EMA shadow swap). The `code_bytes` / `Total submission size` lines reflect the original three-file run; see "Key metrics" above for the post-cleanup numbers.
+- `result.json`, `submission.json` — leaderboard metadata.
+- `requirements.txt` — `brotli` and `sentencepiece` are required at quant-export.
